+++ Runtime/ClientServerWorld/ServerInitializationSystemGroup.cs
@@ -13,11 +13,20 @@ namespace Unity.NetCode
         {
             if (ParentTickSystem != null)
                 ParentTickSystem.RemoveSystemFromUpdateList(this);
+#if !UNITY_DOTSRUNTIME
+            //The playerloop is double buffered. However, becasue the ComponentSystemBase is a managed class
+            //and the StatePtr is set to null, all the DummyWrapper instances will get called but will result in
+            //an early return.
+            ScriptBehaviourUpdateOrder.RemoveWorldFromCurrentPlayerLoop(World);
+#endif
         }
 #endif
     }
 
 #if !UNITY_CLIENT || UNITY_SERVER || UNITY_EDITOR
+#if !UNITY_DOTSRUNTIME
+    [DisableAutoCreation]
+#endif
     [UpdateInGroup(typeof(InitializationSystemGroup))]
     [AlwaysUpdateSystem]
     [UpdateInWorld(TargetWorld.Default)]