+++ Runtime/ClientServerWorld/ServerSimulationSystemGroup.cs
@@ -164,6 +164,9 @@ namespace Unity.NetCode
     }
 
 #if !UNITY_CLIENT || UNITY_SERVER || UNITY_EDITOR
+#if !UNITY_DOTSRUNTIME
+    [DisableAutoCreation]
+#endif
     [AlwaysUpdateSystem]
     [UpdateInWorld(TargetWorld.Default)]
     public class TickServerSimulationSystem : ComponentSystemGroup